+++ addons/im_livechat/__manifest__.py
@@ -68,7 +68,7 @@ Help your customers with this chat, and analyse their feedback.
             'im_livechat/static/src/chat_window/**/*',
             'im_livechat/static/src/composer/**/*',
             'im_livechat/static/src/core/**/*',
-            'im_livechat/static/src/discuss/**/*',
+            'im_livechat/static/src/discuss_app/**/*',
             'im_livechat/static/src/messaging_menu/**/*',
             'im_livechat/static/src/web/**/*',
             'im_livechat/static/src/scss/im_livechat_history.scss',
