+++ addons/im_livechat/static/src/discuss_app/sidebar_patch.js
@@ -1,6 +1,6 @@
 /** @odoo-module */
 
-import { Sidebar } from "@mail/web/discuss/sidebar";
+import { Sidebar } from "@mail/web/discuss_app/sidebar";
 import { patch } from "@web/core/utils/patch";
 
 patch(Sidebar.prototype, "im_livechat", {
