+++ addons/im_livechat/static/src/discuss_app/sidebar_patch.xml
