+++ addons/im_livechat/static/src/discuss_app/thread_icon_patch.xml
