+++ addons/mail/static/src/discuss_app/autoresize_input.xml
