+++ addons/mail/static/src/discuss_app/channel_invitation.xml
