+++ addons/mail/static/src/discuss_app/channel_member_list.xml
