+++ addons/mail/static/src/discuss_app/discuss.scss
