+++ addons/mail/static/src/discuss_app/discuss_client_action.xml
