+++ addons/mail/static/src/discuss_app/im_status.js
