+++ addons/mail/static/src/discuss_app/pinned_messages_panel.js
