+++ addons/mail/static/src/discuss_app/thread_icon.xml
