+++ addons/mail/static/src/web/discuss_app/context_service.js
