+++ addons/mail/static/src/web/discuss_app/discuss_patch.xml
