+++ addons/mail/static/src/web/discuss_app/sidebar.scss
