+++ addons/mail/static/tests/discuss_app/channel_invitation_tests.js
