+++ addons/mail/static/tests/discuss_app/channel_member_list_tests.js
