+++ addons/mail/static/tests/discuss_app/inbox_tests.js
