+++ addons/mail/static/tests/discuss_app/sidebar_tests.js
