+++ addons/website_livechat/static/src/core_ui/thread_patch.js
@@ -1,6 +1,6 @@
 /* @odoo-module */
 
 import { Thread } from "@mail/core_ui/thread";
-import { ImStatus } from "@mail/discuss/im_status";
+import { ImStatus } from "@mail/discuss_app/im_status";
 
 Object.assign(Thread.components, { ImStatus });
